[Bronze V] Title: 이상한 기호, Time: 104 ms, Memory: 14100 KB, Score: 100 point -BaekjoonHub

+++ 백준/Bronze/15964. 이상한 기호/이상한 기호.java
@@ -5,10 +5,10 @@ public class Main {
 	public static void main(String[] args)throws IOException{
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		StringTokenizer st = new StringTokenizer(br.readLine());
-		int A = Integer.parseInt(st.nextToken());
-		int B = Integer.parseInt(st.nextToken());
+		long A = Integer.parseInt(st.nextToken());
+		long B = Integer.parseInt(st.nextToken());
 		
-		int ans = (A+B)*(A-B);
+		long ans = (A+B)*(A-B);
 		
 		System.out.println(ans);
 	}
